requirements.txt
flask
line-bot-sdk
google-generativeai
gunicorn  # 正式環境的伺服器軟體
gevent    # 配合 gunicorn 使用，提升效能